restructure to readers and remove some leftover code
